Solucionado tema controladores

+++ src/controllers/postController.js
@@ -1,48 +1,64 @@
 const { Post } = require('../db/models');
 const { Post_Image } = require('../db/models');
+const { User } = require('../db/models');
+const { Tag } = require('../db/models');
 
 const getPosts = async (req, res) => {
     const data = await Post.find()
-        //.populate("nickName")
-        //.populate("imagenes")
-        //.populate("tags");
-    res.status(200).json(data);
-};
-
-const getPostById = async (req, res) => {
-    const data = await Post.findById(req.params.id)
-        .populate("comentarios")
+        .populate("nickName")
         .populate("imagenes")
         .populate("tags");
     res.status(200).json(data);
 };
 
+const getPostById = async (req, res) => {
+    const data = await Post.findById(req.params.id)
+        .populate("imagenes")
+        .populate("tags");
+    if (!data) return res.status(404).json({ error: 'Post No Encontrado' });
+    res.status(200).json(data);
+};
+
 const createPost = async (req, res) => {
+    const user = await User.findOne({ nickName: req.body.nickName });
+    if (!user) return res.status(404).json({ error: 'NickName No Encontrado' });
+
+    let imagenIds = [];
+    if (req.body.imagenes && Array.isArray(req.body.imagenes)) {
+      const imagenDocs = await Post_Image.find({ url: { $in: req.body.imagenes } });
+      imagenIds = imagenDocs.map(img => img._id);
+    }
+
+    let tagIds = [];
+        if (Array.isArray(req.body.tags)) {
+            const tags = await Tag.find({ tagName: { $in: req.body.tags } });
+            tagIds = tags.map(tag => tag._id);
+        }
+    
     const newPost = await Post.create({
         description: req.body.description,
-        nickName: req.body.nickName,
-        fecha: req.body.fecha
+        nickName: user._id,
+        fecha: req.body.fecha,
+        imagenes: imagenIds,
+        tags: tagIds
     });
-
-    if (req.body.imagenes != null) {
-        for (const i of req.body.imagenes) {
-            await Post_Image.create({ url: i, postId: newPost._id });
-        }
-    }
 
     res.status(201).json(newPost);
 };
 
 const deletePostById = async (req, res) => {
-    const data = await Post.findById(req.params.id).populate("comentarios");
-    await Promise.all(data.comentarios.map(c => c.deleteOne()));
+    const data = await Post.findById(req.params.id)
+    if (!data) return res.status(404).json({ error: 'Post No Encontrado' });
     const removed = await data.deleteOne();
     res.status(200).json(removed);
 };
 
+//Falta probar de aca para abajo ***********************************
+
 const putPostById = async (req, res) => {
     await Post.updateOne({ _id: req.params.id }, req.body);
     const data = await Post.findById(req.params.id);
+    if (!data) return res.status(404).json({ error: 'Post No Encontrado' });
     res.status(201).json(data);
 };
